update Usage module

+++ src/EggFarmSystem/EggFarmSystem/Models/SearchResult.cs
@@ -14,6 +14,6 @@ namespace EggFarmSystem.Models
 
         public List<T> Items { get; set; }
 
-        public long Total { get; set; }
+        public int Total { get; set; }
     }
 }
